+++ app/catalog/handler_test.go
@@ -1,0 +1,1 @@
+package catalog_test
